CompositeLocator refactored to ReferenceComposite
