+++ org.jboss.reddeer.swt/src/org/jboss/reddeer/swt/impl/shell/DefaultShell.java
@@ -4,8 +4,8 @@ import org.eclipse.swtbot.swt.finder.exceptions.WidgetNotFoundException;
 import org.jboss.reddeer.swt.api.Shell;
 import org.jboss.reddeer.swt.condition.ShellWithTextIsActive;
 import org.jboss.reddeer.swt.exception.SWTLayerException;
-import org.jboss.reddeer.swt.locate.CompositeWidget;
-import org.jboss.reddeer.swt.locate.CompositeWidgetLocator;
+import org.jboss.reddeer.swt.reference.ReferenceComposite;
+import org.jboss.reddeer.swt.reference.ReferencedComposite;
 import org.jboss.reddeer.swt.util.Bot;
 import org.jboss.reddeer.swt.wait.WaitUntil;
 
@@ -16,7 +16,7 @@ import org.jboss.reddeer.swt.wait.WaitUntil;
  * @author Jiri Peterka
  *
  */
-public class DefaultShell extends AbstractShell implements Shell, CompositeWidget {
+public class DefaultShell extends AbstractShell implements Shell, ReferencedComposite {
 
 	
 	public DefaultShell(String title) {
@@ -44,8 +44,8 @@ public class DefaultShell extends AbstractShell implements Shell, CompositeWidge
 	}
 
 	@Override
-	public void setCompositeWidget() {
-		CompositeWidgetLocator.setCompositeWidget(null);
+	public void setAsReference() {
+		ReferenceComposite.setComposite(null);
 	}
 }
