+++ org.jboss.reddeer.swt/src/org/jboss/reddeer/swt/reference/ReferencedComposite.java
@@ -1,15 +1,17 @@
-package org.jboss.reddeer.swt.locate;
+package org.jboss.reddeer.swt.reference;
 
 /**
  * Classes implementing this interface can be used
  * as composite widget parent from which children are
  * located
+ * @author Jaroslav Jankovic
+ * @author Jiri Peterka
  */
-public interface CompositeWidget {
+public interface ReferencedComposite {
 
 	/**
 	 * Update the reference of composite widget parent 
 	 */
-	void setCompositeWidget();
+	void setAsReference();
 	
 }
