+++ org.jboss.reddeer.workbench/src/org/jboss/reddeer/workbench/view/WorkbenchPart.java
@@ -5,8 +5,8 @@ import org.eclipse.swtbot.eclipse.finder.widgets.SWTBotWorkbenchPart;
 import org.eclipse.ui.IViewReference;
 import org.jboss.reddeer.swt.condition.WaitCondition;
 import org.jboss.reddeer.swt.exception.Thrower;
-import org.jboss.reddeer.swt.locate.CompositeWidget;
-import org.jboss.reddeer.swt.locate.CompositeWidgetLocator;
+import org.jboss.reddeer.swt.reference.ReferenceComposite;
+import org.jboss.reddeer.swt.reference.ReferencedComposite;
 import org.jboss.reddeer.swt.wait.WaitWhile;
 
 /**
@@ -16,13 +16,13 @@ import org.jboss.reddeer.swt.wait.WaitWhile;
  * @author jjankovi
  * 
  */
-public abstract class WorkbenchPart implements CompositeWidget {
+public abstract class WorkbenchPart implements ReferencedComposite {
 
 	protected final Logger log = Logger.getLogger(this.getClass());
 
 	@Override
-	public void setCompositeWidget() {
-		CompositeWidgetLocator.setCompositeWidget(null);
+	public void setAsReference() {
+		ReferenceComposite.setComposite(null);
 	}
 	
 	public void close() {
